Move functions to load DRP from yaml to its own submodule
