+++ numina/core/__init__.py
@@ -24,7 +24,8 @@ from .dataframe import DataFrame
 from .pipeline import Instrument, Pipeline, InstrumentConfiguration
 from .pipeline import ObservingMode
 from .pipeline import import_object
-from .pipeline import drp_load, init_drp_system
+from .pipeline import init_drp_system
+from .pipelineload import drp_load
 from .requirements import DataProductRequirement
 from .load import RequirementParser
 from .requirements import Parameter, Requirement
